requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==4.2.15
django-filter==24.3
djangorestframework==3.15.2
Faker==28.0.0
importlib_metadata==8.4.0
isort==5.13.2
Markdown==3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
zipp==3.20.1
